# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: SafeDrive AI API
  version: 1.0.0
  description: |
    Complete API specification for SafeDrive AI - an intelligent smartphone-based driver safety system.

    ## Features
    - JWT-based authentication
    - Real-time trip tracking
    - Incident logging (distraction, drowsiness, crash)
    - Video upload with pre-signed S3 URLs
    - User profile management
    - Trip analytics and safety scoring

    ## Authentication
    Most endpoints require Bearer token authentication. Obtain a token via `/auth/login` or `/auth/register`.

    ## Rate Limiting
    - Anonymous: 100 requests/hour
    - Authenticated: 1000 requests/hour

    ## Note on this document
    The `Auth`, `Users`, `Trips`, `Incidents`, and `Videos` sections above
    describe an earlier, illustrative API shape (UUID ids, `/auth/...`
    paths) and have drifted from the current FastAPI implementation, which
    actually uses integer ids and `/users/register`, `/users/login`, etc.
    The `Contacts` and `Alerts` sections were added directly from the real,
    implemented routes and reflect the actual request/response shapes.

  contact:
    name: SafeDrive AI Team
    email: support@safedriveai.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Development
  - url: https://api.safedriveai.com
    description: Production
security:
  - bearerAuth: []
...
paths:
...
  /contacts/:
    post:
      summary: Add a trusted contact
      description: >-
        Saves a phone number (and optional name/email/relationship) that
        will receive an SMS if the authenticated driver is detected as
        continuously distracted for a prolonged period. The contact does
        not need a SafeDrive AI account.
      tags: [Contacts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmergencyContactCreate"
            examples:
              addContact:
                summary: Add a contact
                value:
                  name: "Jamie Driver"
                  phone_number: "+15551234567"
                  relationship: "spouse"
      responses:
        "201":
          description: Contact saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmergencyContact"
        "401":
          $ref: "#/components/responses/Unauthorized"

    get:
      summary: List trusted contacts
      description: Retrieve all trusted contacts saved by the authenticated driver.
      tags: [Contacts]
      responses:
        "200":
          description: List of trusted contacts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EmergencyContact"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /contacts/{contact_id}:
    delete:
      summary: Remove a trusted contact
      description: >-
        Deletes a contact owned by the authenticated driver. Returns 404 if
        the contact does not exist or belongs to a different driver.
      tags: [Contacts]
      parameters:
        - in: path
          name: contact_id
          required: true
          schema:
            type: integer
          description: Contact ID
      responses:
        "204":
          description: Contact deleted
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ============================================================
  # ALERTS
  # ============================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from `/auth/login` or `/auth/register`
  responses:
...
    Unauthorized:
      description: Unauthorized - Missing or invalid token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            noToken:
              summary: No auth token
              value:
                message: "Authorization header missing"
                code: "UNAUTHORIZED"
    NotFound:
      description: Not Found - Resource does not exist
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            resourceNotFound:
              summary: Resource not found
              value:
                message: "Trip not found"
                code: "NOT_FOUND"
  schemas:
...
    EmergencyContactCreate:
      type: object
      required: [name, phone_number]
      properties:
        name:
          type: string
          description: Contact's display name
          example: "Jamie Driver"
        phone_number:
          type: string
          description: Contact's phone number, SMS destination
          example: "+15551234567"
        email:
          type: string
          format: email
          nullable: true
          example: "jamie@example.com"
        relationship:
          type: string
          nullable: true
          description: Optional free-text label
          example: "spouse"
    EmergencyContact:
      allOf:
        - $ref: "#/components/schemas/EmergencyContactCreate"
        - type: object
          required: [id]
          properties:
            id:
              type: integer
              description: Unique contact identifier
              example: 1

    # ============================================================
    # ALERT SCHEMAS
    # ============================================================
...
    Error:
      type: object
      required: [message, code]
      properties:
        message:
          type: string
          description: Human-readable error message
          example: "Invalid email format"
        code:
          type: string
          description: Machine-readable error code
          example: "VALIDATION_ERROR"
...